Implementation of Testing Framework
This PR converts all of the files in the examples directory into go test files. This was done so that in the future, we can have robust unit tests. Instead of "hardcoding" testing variables into each test function, they are now located
in a .env folder. The testing structure was inspired by the Terraform-Provider-IBM repository.

I also removed some unused packages / files.

Changes
 - Created unit tests in "test" directory
    - Deleted examples directory
    - Added ".vscode" directory for test configurations and env variables
    - Added private .env files to .gitignore
 - Removed helpers/env.go. Functions were not being used:
    - EnvFallBack()
 - Removed utils package. Functions were longer used:
    - GetEndpoint()
    - getGCEndpoint()
    - getNGEndpoint()
    - GetNext()
    - Relocated GetPowerEndPoint() from utils to ibmpisession
- Moved logger_test.go to new "test" directory
    - Renamed as ibmpi_logger_test.go
    - Made SanatizeArgs() in ibmpisession an exported function
- Updated vendor directory
- Updated version.go
- Updated go-sdk-core in go.mod to v5.8.2
- Edited README.md